[feat] : #15 FriendRecordGetAdapter inner class seperate
